Updated base complex types with an ODataType property.
https://github.com/microsoftgraph/MSGraph-SDK-Code-Generator/pull/154

Added constructors to initialize OdataType for classes that have an abstract base that is referenced as the type of a property. This fixes #294 (#302)

The result is that all complexTypes have an ODataType property and any complexTypes with an abstract base that is referenced as a property's type will have its odata.type set on it, such as OmaSettingsString.cs.