Add `Object` postfix to type name in `DerivedTypeConverter`
For Graph resources which end with `Request`, the model generator adds
the postfix `Object`. Since we use the ODataType to construct an
instance of the generated class in `DerivedTypeConverter`, we actually
attempt to create instances of the wrong type. To fix this, we also add
the postfix to the type name in `DerivedTypeConverter`, before
attempting to create instances of the type.

Fixes msgraph-sdk-dotnet#1449.